% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_run.R
\name{loss_run_all_evals}
\alias{loss_run_all_evals}
\title{loss_run_all_evals}
\usage{
loss_run_all_evals(trans_, devt_months = NULL, merge = FALSE)
}
\arguments{
\item{trans_}{transactional claims data}

\item{devt_months}{optinal character vector (numeric) of months to inlcude}

\item{merge}{logical - merge into data.frame or keep as list?}
}
\value{
either a list or merged df
}
\description{
wrapper around loss_run
}
