% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pull_unique}
\alias{pull_unique}
\title{Pull Unique Values from a dataframe}
\usage{
pull_unique(df, var)
}
\arguments{
\item{df}{Dataframe}

\item{var}{Quoted named of variable}
}
\value{
Character vector of unique, sorted values from specified column
}
\description{
Pull Unique Values from a dataframe
}
