% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lossdevt-package.R
\docType{package}
\name{lossdevt-package}
\alias{lossdevt}
\alias{lossdevt-package}
\title{lossdevt: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jimbrig.github.io/lossdevt}
}

}
\author{
\strong{Maintainer}: Jimmy Briggs \email{jimmy.briggs@jimbrig.com.} (\href{https://orcid.org/0000-0002-7489-8787}{ORCID})

}
\keyword{internal}
